Re organized script into multiple files
